pyproject.toml
[project]
name = "battle_test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.9",
  "rich>=13.0",
  "torch",
  "openai-whisper @ git+https://github.com/openai/whisper.git",
  "numpy==1.26.4"
]

[project.scripts]
main = "battle_test.main:main"


# =============================================================================
# UV Settings
# =============================================================================

[tool.uv]
resolution = "lowest-direct"
dev-dependencies = [
  # <br>
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
]





# =============================================================================
# Build System
# =============================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

# =============================================================================
# Type Checking
# =============================================================================

[tool.pyright]
include = []
exclude = ["**/__pycache__"]
venvPath = "."
venv = ".venv"

reportMissingImports = "error"
reportMissingTypeStubs = false

pythonVersion = "3.12"
pythonPlatform = "Linux"

executionEnvironments = [{ root = "src" }]

# =============================================================================
# Testing
# =============================================================================

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-xvs"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
